refactor: websocket 관련 service 제거하고 gateway를 늘려거 logic 해결하도록 수정
